+++ CHANGELOG.md
@@ -7,11 +7,18 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased [v0.4.0] - (2021-XX-XX)
 ### Fixes
+
 ### Breaking changes
+
+- The CLI API has changed; now it must be invoked by specifying the subcommand:
+  ```
+  $ terravalet rename -plan=PLAN -up=UP.sh -down=DOWN.sh
+  ```
+
 ### Changes
 ### New
 
-## Unreleased [v0.3.0] - (2020-12-11)
+## [v0.3.0] - (2020-12-11)
 
 ### New